refactor: split out methods
