+++ loadPkg.js
@@ -1,0 +1,19 @@
+const fs = require('fs')
+const path = require('path')
+const readFile = require('util').promisify(fs.readFile)
+
+async function loadPkg(pkg) {
+  if (pkg === 'vue') {
+    const dir = path.dirname(require.resolve('vue'))
+    const filepath = path.join(dir, 'vue.esm.browser.js')
+    return readFile(filepath)
+  }
+  else {
+    // TODO
+    // check if the package has a browser es module that can be used
+    // otherwise bundle it with rollup on the fly?
+    throw new Error('npm imports support are not ready yet.')
+  }
+}
+
+exports.loadPkg = loadPkg
